tests done, pino logger, etc

+++ src/application/users/services/user.service.mjs
@@ -1,6 +1,8 @@
+import mongoose from 'mongoose';
 import User from '../../../domain/users/entities/user.entity.mjs';
+import customError from '../../../handlers/customError.mjs';
 
-export class UserService {
+export default class UserService {
   userRepository;
   constructor(userRepository) {
     this.userRepository = userRepository;
@@ -12,7 +14,15 @@ export class UserService {
   }
 
   async getOneById(id) {
-    const user = await this.userRepository.getOneById(id);
+    if (!mongoose.isValidObjectId(id)) {
+      return customError(400, 'Invalid ID format.');
+    }
+
+    const user = await this.userRepository.getOneById(mongoose.Types.ObjectId(id));
+
+    if (!user) {
+      return customError(404, 'User not found.');
+    }
     return user;
   }
 
@@ -21,7 +31,7 @@ export class UserService {
     const { errors, isValid } = newUser.isValid();
 
     if (!isValid) {
-      throw new Error(errors.join(" "));
+      return customError(400, errors.join(" "));
     }
 
     const persistedUser = await this.userRepository.create(newUser);
@@ -29,11 +39,28 @@ export class UserService {
   }
 
   async update(id, data) {
+    if (!mongoose.isValidObjectId(id)) {
+      return customError(400, 'Invalid ID format.');
+    }
+
+    const userExists = await this.userRepository.getOneById(mongoose.Types.ObjectId(id));
+
+    if (!userExists) {
+      return customError(404, 'User not found.');
+    }
     const user = await this.userRepository.update(id, data);
     return user;
   }
 
   async delete(id) {
+    if (!mongoose.isValidObjectId(id)) {
+      return customError(400, 'Invalid ID format.');
+    }
+
+    const userExists = await this.userRepository.getOneById(mongoose.Types.ObjectId(id));
+    if (!userExists) {
+      return customError(404, 'User not found.');
+    }
     await this.userRepository.delete(id);
   }
 }
